add element helpers
